+++ crates/nostr-relay-pool/src/pool/constants.rs
@@ -4,7 +4,5 @@
 
 //! Constants
 
-pub(super) const MAX_CONNECTING_CHUNK: usize = 100;
-
 /// Relay Pool default notification channel size
 pub const DEFAULT_NOTIFICATION_CHANNEL_SIZE: usize = 4096;