+++ crates/nostr-sdk/src/client/mod.rs
@@ -493,13 +493,27 @@ impl Client {
     }
 
     /// Connect to a previously added relay
+    ///
+    /// Check [`RelayPool::connect_relay`] docs to learn more.
     #[inline]
     pub async fn connect_relay<U>(&self, url: U) -> Result<(), Error>
     where
         U: TryIntoUrl,
         pool::Error: From<<U as TryIntoUrl>::Err>,
     {
-        Ok(self.pool.connect_relay(url, None).await?)
+        Ok(self.pool.connect_relay(url).await?)
+    }
+
+    /// Try to connect to a previously added relay
+    ///
+    /// For further details, see the documentation of [`RelayPool::try_connect_relay`].
+    #[inline]
+    pub async fn try_connect_relay<U>(&self, url: U, timeout: Duration) -> Result<(), Error>
+    where
+        U: TryIntoUrl,
+        pool::Error: From<<U as TryIntoUrl>::Err>,
+    {
+        Ok(self.pool.try_connect_relay(url, timeout).await?)
     }
 
     /// Disconnect relay
@@ -515,16 +529,29 @@ impl Client {
     /// Connect to all added relays
     #[inline]
     pub async fn connect(&self) {
-        self.pool.connect(None).await;
+        self.pool.connect().await;
+    }
+
+    /// Try to establish a connection with the relays.
+    ///
+    /// Attempts to establish a connection without spawning the connection task if it fails.
+    /// This means that if the connection fails, no automatic retries are scheduled.
+    /// Use [`Client::connect`] if you want to immediately spawn a connection task,
+    /// regardless of whether the initial connection succeeds.
+    ///
+    /// For further details, see the documentation of [`RelayPool::try_connect`].
+    #[inline]
+    pub async fn try_connect(&self, timeout: Duration) -> Output<()> {
+        self.pool.try_connect(timeout).await
     }
 
     /// Connect to all added relays
     ///
     /// Try to connect to the relays and wait for them to be connected at most for the specified `timeout`.
     /// The code continues if the `timeout` is reached or if all relays connect.
-    #[inline]
+    #[deprecated(since = "0.39.0", note = "Use `try_connect` instead")]
     pub async fn connect_with_timeout(&self, timeout: Duration) {
-        self.pool.connect(Some(timeout)).await
+        self.pool.try_connect(timeout).await;
     }
 
     /// Disconnect from all relays